With this commit the membership addendum is fully up to date with what was voted in. The main bylaws should theoretically be correct, but they have not been fully reviewed yet.
For reference here are the meeting minutes for when the membership
addendum was voted in: https://wiki.hive13.org/view/Nov_27,_2012

I also removed the author tag since git history should show that. I
removed the date as well since the addendum will be attached to the
bylaws which will have the correct date on them.